% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_number.R
\name{pretty_number}
\alias{pretty_number}
\title{Format numbers without scientific notation}
\usage{
pretty_number(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
A character string.
}
\description{
Ensures clean, publication-ready formatting.
}
